feat(Misc): rework module
