+++ src/NpcBeastmaster_loader.cpp
